isolate test runner and code modules into separate python modules
