+++ systemtests/nidcpower_codemodules.py
@@ -1,20 +1,8 @@
-import pytest
 import nidcpower
 import nitsm.codemoduleapi
 from nitsm.codemoduleapi import SemiconductorModuleContext
 
 OPTIONS = {"Simulate": True, "DriverSetup": {"Model": "4141", "BoardType": "PXIe"}}
-
-
-@pytest.mark.skip("Can't enable this test until nidcpower python supports channel expansion.")
-@pytest.mark.sequence_file("nidcpower.seq")
-def test_nidcpower(system_test_runner):
-    assert system_test_runner.run()
-
-
-@pytest.mark.sequence_file("nidcpower_legacy.seq")
-def test_nidcpower_legacy(system_test_runner):
-    assert system_test_runner.run()
 
 
 @nitsm.codemoduleapi.code_module
